requirements.txt
certifi==2019.9.11
Django==2.2.7
django-cors-headers==3.1.1
django-dotenv==1.4.2
django-filter==2.2.0
django-redis==4.10.0
djangorestframework==3.10.3
drf-extensions==0.5.0
gunicorn==20.0.0
mysqlclient==1.4.5
PyJWT==1.7.1
pytz==2019.3
redis==3.3.11
sentry-sdk==0.13.2
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
